Changes percent_ram models data type from int to float

+++ src/emc/core/models.py
@@ -1,5 +1,6 @@
 from dataclasses import dataclass
 
+
 @dataclass
 class OSInfo:
     usr_name: str
@@ -7,20 +8,24 @@ class OSInfo:
     os_ver: str
     os_sys_type: str
 
+
 @dataclass
 class CPUInfo:
     cpu_name: str
     cpu_cores: int
     cpu_threads: int
 
+
 @dataclass
 class MEMInfo:
     total_ram: float
     avail_ram: float
     used_ram: float
-    percent_ram: int
+    percent_ram: float
+
 
 @dataclass
 class DISKInfo:
     sum_os_storage: float
     free_os_storage: float
+
